feat: 요일별 웹툰 리스트 화면에 사용할 itemDecoration 에서 Margin 대신 resourced에서 계산된 pixel을 사용하도록 변경
