+++ app/src/main/kotlin/kr/tooni/tooni/utils/GridItemDecoration.kt
@@ -1,0 +1,30 @@
+/*
+ * Created by Leo on 2021. 05. 23 ..
+ */
+package kr.tooni.tooni.utils
+
+import android.graphics.Rect
+import android.view.View
+import androidx.recyclerview.widget.RecyclerView
+
+class GridItemDecoration(
+    private val spanCount: Int,
+    private val spacing: Int
+) : RecyclerView.ItemDecoration() {
+    
+    override fun getItemOffsets(
+        outRect: Rect,
+        view: View,
+        parent: RecyclerView,
+        state: RecyclerView.State
+    ) {
+        val position = parent.getChildAdapterPosition(view)
+        val column = position % spanCount
+    
+        outRect.left = column * spacing / spanCount
+        outRect.right = spacing - (column + 1) * spacing / spanCount
+        if (position >= spanCount) {
+            outRect.top = spacing
+        }
+    }
+}
