+++ app/src/main/res/layout/item_webtoon.xml
@@ -29,7 +29,7 @@
             app:cardBackgroundColor="@color/white"
             app:cardCornerRadius="@dimen/space_5"
             app:cardElevation="@dimen/space_0"
-            app:layout_constraintDimensionRatio="h, 1:1"
+            app:layout_constraintDimensionRatio="1"
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toTopOf="parent">
